also from twaugh:
I found some input that goes wrong with the synop.xsl we have in CVS:

<group choice="opt">
  <arg>-o <replaceable>FILE</replaceable></arg>
  <arg>--output=<replaceable>FILE</replaceable></arg>
</group>

It gets rendered as (with *bold* and _italic_):

[*-o FILE* | *--output=FILE*]

The desired markup should look like:
The following macro does the trick:

[\fB-o \fIFILE\fR\fR | \fB--output=\fIFILE\fR\fR]

The trouble is that the named template 'bold' uses value-of, and so
strips <replaceable> of its significance.

Another thing I found is that the arg/replaceable template is
superfluous altogether: db2man.xsl has a 'replaceable' template which
does the same thing.

Here is a patch to make those two modifications.

NOTE TO SELF: must try to fix bold template so we can use it everywhere