Trademark symbol handling made consistent with handling of same in HTML stylesheets. (closes #1218286; thanks to Mauritz Jeanson for reporting the problem)
Prior to this change, if you processed a doc that contained no
value for the Class attribute on the Trademark element, the HTML
stylesheets would default to rendering a superscript TM symbol
after the Trademark contents, but the FO stylesheets would render
nothing. This change alters the FO handing of Trademark such that
it is now identical to the HTML handling.