Made some minor changes to get deal with a couple corner cases.
If somebody (say, Mike) checks in the VERSION file twice with the
same change description, it's possible for the merged changelog to
end with up something like:

  * VERSION, VERSION: Version 1.69.1 released.

The existing version of mergechangelogs got confused by that, so I
added a regex match to catch instances like that.

Also, in running the snapshot builds, with the way I have the
snapshot build currently set up, it's useful for the build to just
automatically assume we are at version X.XX.1 if the previous
version was X.XX.0.
----------------------------------------------------------------------