Drastically reworked all of the XPath expressions used in refentry metadata gathering -- completely removed $parentinfo and turned $info into a set of nodes that includes the *info contents of the Refentry plus the *info contents all all of its ancestor elements. The basic XPath expression now used throughout is (using the example of checking for a date):
  (($info[//date])[last()]/date)[1].

That selects the "last" *info/date date in document order -- that
is, the one eitther on the Refentry itself or on the closest
ancestor to the Refentry.

It's likely this change may break some things; may need to pick
up some pieces later.

Also, changed the default value for the man.th.extra2.max.length
from 40 to 30.