Added the man.output.in.separate.dir, man.output.base.dir, and man.output.subdirs.enabled parameters.
The man.output.base.dir parameter specifies the base directory
into which man-page files are output. The
man.output.subdirs.enabled parameter controls whether the files
are output in subdirectories within the base directory.

The values of the man.output.base.dir and
man.output.subdirs.enabled parameters are used only if the value
of man.output.in.separate.dir parameter is non-zero. If the value
of man.output.in.separate.dir is zero, man-page files are not
output in a separate directory.