Get all metadata from resource referenced by merge element
When a <merge> element uses a resourceref attribute to point to a
resource, the stylesheet collects the <title>, <titleabbrev> and
<subtitle> elements from the top level of the resource (if they
exist), concatenates them with the content of the <info> element
from the top level of the resource (if that exists), and outputs
them all to the <info> element in the assembled document.

The warning that the referenced file must contain an info element
is removed.