fix Muenchian grouping in w/type grouping-problems
When @type was added to DocBook, and thus to DocBook XSL 's index-generation (autoidx.xsl), not enough thought was given on how to incorporate this new factor into the "Muenchian grouping" code that had formed the organizing mechanism for the indexes' contents. Turns out, just tossing in '[&scope;] as a predicate, the way it was done, was not all that needed doing.

In fact, all predicate-tests following this pattern:

  [count(.|key('primary',                        &primary;)[&scope;][1]) = 1]

or even this pattern:

[generate-id(key('primary-section', concat($key, &sep;, &section.id;))[&scope;][1])]

need changing to this pattern:

[generate-id(key('primary-section', concat($key, &sep;, &section.id;))[&scope;])]