Extension function instead of Saxon6 <saxon:script>
`<saxon:script>` is a synonym of `<xsl:script>`. The advice for `<xsl:script>` is that "it's simpler to avoid using this element altogether, and instead to use a namespace URI that identifies the extension function directly". (http://saxon.sourceforge.net/saxon6.5.5/xsl-elements.html#xsl:script)

My real reason for the change is that `function-available('s6hl:highlight')` now actually works, instead of the transformation failing at the `<saxon:script>` when the extension is not available.