Corrected and updated nb and nn locales
Corrected several strings and capitalisation after validating
the strings using both HTML and FO/PDF output.  Based on input from
Kolbjørn Stuestøl, Karl Ove Hufthammer and the Norwegian translators
mailing list.

As Norwegian book colophon pages do not use the 'copyright' phrase,
just the c in a circle, make that phrase empty for nb and nn.

Rename and sync section translation for nb and nn to avoid
name collision with part.  Updated several formatting
templates based on Norwegian language rules.

Added new strings and reordered the nb and nn locales to match the
order used in the en locale.  This make it easier to discover
additions to the 'en' locale in the future, as well as comparing
the nb and nn locales to each other.   Removed duplicate entires
added by mistake.